+++ crates/matrix-sdk-indexeddb/src/state_store/mod.rs
@@ -1946,44 +1946,6 @@ impl_state_store!({
         )
     }
 
-    async fn upsert_thread_subscription(
-        &self,
-        room: &RoomId,
-        thread_id: &EventId,
-        subscription: StoredThreadSubscription,
-    ) -> Result<()> {
-        let encoded_key = self.encode_key(keys::THREAD_SUBSCRIPTIONS, (room, thread_id));
-
-        let tx = self
-            .inner
-            .transaction(keys::THREAD_SUBSCRIPTIONS)
-            .with_mode(TransactionMode::Readwrite)
-            .build()?;
-        let obj = tx.object_store(keys::THREAD_SUBSCRIPTIONS)?;
-
-        let mut new = PersistedThreadSubscription::from(subscription);
-
-        // See if there's a previous subscription.
-        if let Some(previous_value) = obj.get(&encoded_key).await? {
-            let previous: PersistedThreadSubscription = self.deserialize_value(&previous_value)?;
-
-            // If the previous status is the same as the new one, don't do anything.
-            if new == previous {
-                return Ok(());
-            }
-            if !compare_thread_subscription_bump_stamps(previous.bump_stamp, &mut new.bump_stamp) {
-                return Ok(());
-            }
-        }
-
-        let serialized_value = self.serialize_value(&new);
-        obj.put(&serialized_value?).with_key(encoded_key).build()?;
-
-        tx.commit().await?;
-
-        Ok(())
-    }
-
     async fn upsert_thread_subscriptions(
         &self,
         updates: Vec<(&RoomId, &EventId, StoredThreadSubscription)>,